requirements.txt
Django~=2.0.6
django-ckeditor
